refactor test imports to align with new CSV data package structure
