+++ src/test/kotlin/data/csv_parser/UserCsvParserTest.kt
@@ -1,9 +1,11 @@
 package data.csv_parser
 
 import com.google.common.truth.Truth
-import data.dto.UserDto
-import data.util.ADMIN
-import data.util.MATE
+import data.csv_data.csv_parser.CsvHandler
+import data.csv_data.csv_parser.UserCsvParser
+import data.csv_data.util.ADMIN
+import data.csv_data.util.MATE
+import data.mongodb_data.dto.UserDto
 import io.mockk.every
 import io.mockk.mockk
 import org.junit.jupiter.api.BeforeEach
@@ -113,9 +115,9 @@ class UserCsvParserTest {
     private fun getCsvLines(): List<String> {
         return listOf(
             "id,name,hashedPassword,role,projectIds",
-            "82e16049-a9fb-4f69-b6f7-3336b68f2ae4,Aiman,3336b68f2ae4,${ADMIN},[]",
-            "045e2ef6-a9f8-43d9-9c33-da8cf3a0ff2f,Zeyad,09sd23od3skd,${ADMIN},[82e16049-a9fb-4f69-b6f7-3336b68f2ae4]",
-            "07f641d4-077e-4f08-978d-3b6c9587f4bf,Yaser,wk5dr98sd6dd,${MATE},[82e16049-a9fb-4f69-b6f7-3336b68f2ae4, 82e16049-a9fb-4f69-b6f7-3336b68f2ae4]",
+            "82e16049-a9fb-4f69-b6f7-3336b68f2ae4,Aiman,3336b68f2ae4,$ADMIN,[]",
+            "045e2ef6-a9f8-43d9-9c33-da8cf3a0ff2f,Zeyad,09sd23od3skd,$ADMIN,[82e16049-a9fb-4f69-b6f7-3336b68f2ae4]",
+            "07f641d4-077e-4f08-978d-3b6c9587f4bf,Yaser,wk5dr98sd6dd,$MATE,[82e16049-a9fb-4f69-b6f7-3336b68f2ae4, 82e16049-a9fb-4f69-b6f7-3336b68f2ae4]",
         )
     }
