+++ src/test/kotlin/data/repositories/UserRepositoryImplTest.kt
@@ -1,8 +1,9 @@
 package data.repositories
 
-import data.csv_parser.CsvHandler
-import data.csv_parser.UserCsvParser
-import data.datasource.UserDataSource
+import data.csv_data.csv_parser.CsvHandler
+import data.csv_data.csv_parser.UserCsvParser
+import data.csv_data.datasource.UserDataSource
+import data.csv_data.repositories.UserRepositoryImpl
 import io.mockk.mockk
 import logic.models.User
 import logic.models.UserRole
